set up for the design for the PersonalityDetailView
