+++ CharacterCreator/Views/DetailViews/BasicDetailView.swift
