+++ CharacterCreator/Views/DetailViews/BasicDetailView.xib
