+++ CharacterCreator/Views/DetailViews/PersonalityDetailView.swift
@@ -1,0 +1,28 @@
+//
+//  PersonalityDetailView.swift
+//  CharacterCreator
+//
+//  Created by Ike Mattice on 4/4/19.
+//  Copyright © 2019 Ike Mattice. All rights reserved.
+//
+
+import UIKit
+
+@IBDesignable
+class PersonalityDetailView: XibView {
+
+	required init?(coder aDecoder: NSCoder) {
+		super.init(coder: aDecoder)
+
+		config()
+	}
+	override init(frame: CGRect) {
+		super.init(frame: frame)
+
+		config()
+	}
+
+	func config() {
+
+	}
+}
